+++ src/timber_design/ghpython/components/CT_Install/code.py
@@ -1,6 +1,6 @@
 """This is the ONLY component that should have this r: specifier! to prevent rhino from forever re-installing everything."""
 
-# r: timber_design==0.3.0, compas_timber==2.2.0
+# r: timber_design==0.3.1, compas_timber==2.2.0
 import compas
 import compas_timber