+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+* Added `CT: PlateStocks` and `CT: NestPlates` GH components.
+* Added `Nesting` input parameter to `CT_BTLx` GH component.
+
 ### Changed
 
 ### Removed