+++ CHANGELOG.md
@@ -26,6 +26,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * `max_cluster_size` parameter from `get_clusters_from_model`.
 
 
+## [0.3.1] 2026-07-09
+
+### Added
+
+### Changed
+
+* Fixed `CT: L Topological Joint Rules`, `CT: Direct Joint Rules`, `CT: Joint Rules From List`, and `CT: Category Joints Rules` — selecting `LButtJoint` crashed the component.
+
+### Removed
+
+
 ## [0.3.0] 2026-07-02
 
 ### Added